copy cleint

+++ sever/service/query.js
@@ -11,24 +11,29 @@ export default async function getQuery(table, action, fieldsUpdate) {
             query = `SELECT * FROM ${DB_NAME}.${table}`;
             break;
         case 'getById':
-            query = `SELECT * FROM ${DB_NAME}.${table} WHERE id = ?`;
+            query = `SELECT * FROM ${DB_NAME}.${table} WHERE id = ? AND status=true`;
             break;
         case 'create':
             const queryFields = `DESC ${DB_NAME}.${table}`;
             const fieldsData = await executeQuery(queryFields);
             const fields = fieldsData.map(row => row.Field);
-            const newFields = fields.filter(field => field !== 'id');
+            const newFields = fields.filter(field => field !== 'id' && field !== 'status');
             const questionMarks = Array(newFields.length).fill('?');
             query = `INSERT INTO ${DB_NAME}.${table} (${newFields.join(', ')}) VALUES (${questionMarks.join(', ')})`;
-            // INSERT INTO DB_PROJECT.users (username, email, address, phone) VALUES (?, ?, ?, ?)
+            // for example what would happen
+            //INSERT INTO DB_PROJECT.users (username, email, address, phone, status) VALUES (?, ?, ?, ?,?)
             break;
         case 'update':
             const fieldsUpdateStr = Object.keys(fieldsUpdate).map(field => `${field} = ?`).join(', ');
             query = `UPDATE ${DB_NAME}.${table} SET ${fieldsUpdateStr} WHERE id = ?`;
+            // for example what would happen
             // UPDATE DB_PROJECT.users SET  email = ?, address = ? WHERE id = ?
             break;
         case 'delete':
-            query = `DELETE FROM ${DB_NAME}.${table} WHERE id = ?`;
+            if(table  == 'users')
+                query=`UPDATE ${DB_NAME}.${table} SET status = false WHERE id = ?`
+            else
+                query = `DELETE FROM ${DB_NAME}.${table} WHERE id = ?`;
             break;
         default:
             throw new Error('Invalid action');
